Removed map.cpp, renamed map_static.cpp to map.cpp
