+++ unit_tests/map.cpp
@@ -17,49 +17,239 @@
 *
 * See COPYRIGHT.txt for copyright notices and details.
 */
-#include <vector>
+#include <atomic>
 
 #include <gtest/gtest.h>
 
 #include "map.h"
 #include "common/polymorphic_execution.h"
 
+#include "supported_executions.h"
+
 using namespace std;
+using namespace grppi;
 
-TEST(map, next_value)
+template <typename T>
+class map_test : public ::testing::Test {
+public:
+  T execution_;
+  polymorphic_execution poly_execution_ = 
+    make_polymorphic_execution<T>();
+
+  // Vectors
+  vector<int> v{};
+  vector<int> v2{};
+  vector<int> v3{};
+  vector<int> w{};
+  vector<int> expected{};
+
+  // Invocation counter
+  std::atomic<int> invocations{0};
+
+  void setup_empty() {
+  }
+
+  void check_empty() {
+    ASSERT_EQ(0, invocations); // Functor was not invoked
+  }
+
+  void setup_single() {
+    v = vector<int>{42};
+    w = vector<int>{99};
+  }
+
+  void check_single() {
+    EXPECT_EQ(1, this->invocations); // one invocation
+    EXPECT_EQ(84, this->w[0]);
+  }
+
+  void setup_single_ary() {
+    v = vector<int>{11};
+    v2 = vector<int>{22};
+    v3 = vector<int>{33};
+    w = vector<int>{99};
+  }
+
+  void check_single_ary() {
+    EXPECT_EQ(1, this->invocations); // one invocation
+    EXPECT_EQ(66, this->w[0]);
+  }
+
+  void setup_multiple() {
+    v = vector<int>{1,2,3,4,5};
+    w = vector<int>(5);
+    expected = vector<int>{2,4,6,8,10};
+  }
+
+  void check_multiple() {
+    EXPECT_EQ(5, this->invocations); // five invocations
+    EXPECT_TRUE(equal(begin(this->expected), end(this->expected), begin(this->w)));
+  }
+
+  void setup_multiple_ary() {
+    v = vector<int>{1,2,3,4,5};
+    v2 = vector<int>{2,4,6,8,10};
+    v3 = vector<int>{10,10,10,10,10};
+    w = vector<int>(5);
+    expected = vector<int>{13,16,19,22,25};
+  }
+
+  void check_multiple_ary() {
+    EXPECT_EQ(5, this->invocations); // five invocations
+    EXPECT_TRUE(equal(begin(this->expected), end(this->expected), begin(this->w)));
+  }
+
+};
+
+// Test for execution policies defined in supported_executions.h
+TYPED_TEST_CASE(map_test, executions);
+
+TYPED_TEST(map_test, static_empty)
 {
-  auto e = grppi::make_polymorphic_execution<grppi::sequential_execution>();
-  vector<int> v{1,2,3,4,5};
-  vector<int> w(5);
-  grppi::map(e, begin(v), end(v), begin(w), [](int x) { return x+1; });
-
-  vector<int> res{2, 3, 4, 5, 6};
-  EXPECT_EQ(res, w);
+  this->setup_empty();
+  grppi::map(this->execution_, begin(this->v), end(this->v), begin(this->w),
+    [this](int x) { 
+      this->invocations++; 
+      return x; 
+    }
+  );
+  this->check_empty();
 }
 
-TEST(map, add2_vec)
+TYPED_TEST(map_test, poly_empty)
 {
-  auto e = grppi::make_polymorphic_execution<grppi::sequential_execution>();
-  vector<int> v{1,2,3,4,5};
-  vector<int> w{2,4,6,8,10};
-  vector<int> r(5);
-  grppi::map(e, begin(v), end(v), begin(r), [](int x, int y) { return x+y; }, begin(w));
-
-  vector<int> res{3, 6, 9, 12, 15};
-  EXPECT_EQ(res, r);
+  this->setup_empty();
+  grppi::map(this->poly_execution_, begin(this->v), end(this->v), begin(this->w),
+    [this](int x) { 
+      this->invocations++; 
+      return x; 
+    }
+  );
+  this->check_empty();
 }
 
-TEST(map, add3_vec)
+TYPED_TEST(map_test, static_empty_ary)
 {
-  auto e = grppi::make_polymorphic_execution<grppi::sequential_execution>();
-  vector<int> v{1,2,3,4,5};
-  vector<int> w1{2,4,6,8,10};
-  vector<int> w2{10, 20, 30, 40, 50};
-  vector<int> r(5);
-  grppi::map(e, begin(v), end(v), begin(r), 
-             [](int x, int y, int z) { return x+y+z; },
-             begin(w1), begin(w2));
+  this->setup_empty();
+  grppi::map(this->execution_, begin(this->v), end(this->v), begin(this->w),
+    [this](int x, int y, int z) { 
+      this->invocations++; 
+      return x+y+z; 
+    },
+    begin(this->v2), begin(this->v3)
+  );
+  this->check_empty();
+}
 
-  vector<int> res{13, 26, 39, 52, 65};
-  EXPECT_EQ(res, r);
+TYPED_TEST(map_test, poly_empty_ary)
+{
+  this->setup_empty();
+  grppi::map(this->poly_execution_, begin(this->v), end(this->v), begin(this->w),
+    [this](int x, int y, int z) { 
+      this->invocations++; 
+      return x+y+z; 
+    },
+    begin(this->v2), begin(this->v3)
+  );
+  this->check_empty();
+}
+
+TYPED_TEST(map_test, static_single)
+{
+  this->setup_single();
+  grppi::map(this->execution_, begin(this->v), end(this->v), begin(this->w),
+    [this](int i) {
+      this->invocations++; 
+      return i*2; 
+    }
+  );
+  this->check_single();
+}
+
+TYPED_TEST(map_test, poly_single)
+{
+  this->setup_single();
+  grppi::map(this->poly_execution_, begin(this->v), end(this->v), begin(this->w),
+    [this](int i) {
+      this->invocations++; 
+      return i*2; 
+    }
+  );
+  this->check_single();
+}
+
+TYPED_TEST(map_test, static_single_ary)
+{
+  this->setup_single_ary();
+  grppi::map(this->execution_, begin(this->v), end(this->v), begin(this->w),
+    [this](int x, int y, int z) {
+      this->invocations++; 
+      return x+y+z; 
+    },
+    begin(this->v2), begin(this->v3)
+  );
+  this->check_single_ary();
+}
+
+TYPED_TEST(map_test, poly_single_ary)
+{
+  this->setup_single_ary();
+  grppi::map(this->execution_, begin(this->v), end(this->v), begin(this->w),
+    [this](int x, int y, int z) {
+      this->invocations++; 
+      return x+y+z; 
+    },
+    begin(this->v2), begin(this->v3)
+  );
+  this->check_single_ary();
+}
+
+TYPED_TEST(map_test, static_multiple)
+{
+  this->setup_multiple();
+  grppi::map(this->execution_, begin(this->v), end(this->v), begin(this->w),
+    [this](int i) {
+      this->invocations++; 
+      return i*2; 
+    }
+  );
+  this->check_multiple();
+}
+
+TYPED_TEST(map_test, poly_multiple)
+{
+  this->setup_multiple();
+  grppi::map(this->poly_execution_, begin(this->v), end(this->v), begin(this->w),
+    [this](int i) {
+      this->invocations++; 
+      return i*2; 
+    }
+  );
+  this->check_multiple();
+}
+
+TYPED_TEST(map_test, static_multiple_ary)
+{
+  this->setup_multiple_ary();
+  grppi::map(this->execution_, begin(this->v), end(this->v), begin(this->w),
+    [this](int x, int y, int z) {
+      this->invocations++; 
+      return x+y+z; 
+    },
+    begin(this->v2), begin(this->v3)
+  );
+  this->check_multiple_ary();
+}
+
+TYPED_TEST(map_test, poly_multiple_ary)
+{
+  this->setup_multiple_ary();
+  grppi::map(this->poly_execution_, begin(this->v), end(this->v), begin(this->w),
+    [this](int x, int y, int z) {
+      this->invocations++; 
+      return x+y+z; 
+    },
+    begin(this->v2), begin(this->v3)
+  );
+  this->check_multiple_ary();
 }
